requirements.txt
pandas==2.3.2
numpy==2.3.2
fastapi==0.114.0
uvicorn==0.23.2
requests==2.32.0
httpx==0.26.0
python-dateutil==2.8.2
textblob==0.17.1
nltk==3.8.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-generativeai==0.8.5
streamlit==1.36.0
openai
asyncpraw
psycopg2-binary